requirements-test.txt
-e .
coverage==7.6.4
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
pytest-subtests==0.13.1
